mips64n32: Fix the CALL_FN_ macros by using long long everywhere
This changes the type of the argvec and res to be equal to register size
on both mips64n32 and mips64.

It's not clear that passing 64bit arguments is useful on n32. An
alternative solution would be to have an #ifdef PLAT_mips64n32_linux
which redefines all the macros and changes asm offsets to multiples of 4
instead of 8.